requirements.txt
torch
torchvision
torchaudio
numpy
pandas
scikit-learn
seaborn
sentence-transformers
pypdf
jq
uuid
langchain
langchainhub
langchain-nomic
langchain-community
langchain-experimental
langchain-ollama
langchain-huggingface
langchain_core
langchain_openai
langchain_text_splitters==0.2.2
langgraph
langsmith
openai==1.35.15
SQLAlchemy==2.0.31
streamlit==1.36.0
duckduckgo-search==6.2.1
fastembed==0.1.3
docarray==0.40.0
validators==0.33.0
python-dotenv
faiss-cpu
flashrank
transformers
tokenizers